Add event listener to each box to change color to black
The event listener is added to each box with a forEach loop cycling
through an array of div elements from the boxes class. The event
listener calls on the changeColor function upon the mouse hovering over
the div, which changes background color to black. The array was created
from the nodeList of the boxes class.